refactor(envelope): split outer.rs into pure types + async MLS ops (#1446)
Prerequisite for scp-protocol crate extraction. Separates pure sync
protocol types (OuterEnvelope, serialization, parsing) from async MLS
operations (seal_envelope, open_envelope) using directory promotion,
consistent with the inner.rs split pattern.
